Moved tree series mixin as TreeComposition.
